requirements.txt
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
astroid==3.3.5
attrs==24.2.0
bandit==1.8.0
beautifulsoup4==4.12.3
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cssselect2==0.7.0
dataclasses-json==0.6.7
dill==0.3.9
distro==1.9.0
Django==5.1.3
djangorestframework==3.15.2
eval_type_backport==0.2.0
faiss-cpu==1.9.0.post1
filelock==3.16.1
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
gpt-index==0.8.42
greenlet==3.1.1
grpcio==1.68.0
grpcio-tools==1.68.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.26.2
hyperframe==6.0.1
idna==3.10
isort==5.13.2
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.4.2
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-huggingface==0.1.2
langchain-ollama==0.2.0
langchain-openai==0.2.10
langchain-text-splitters==0.3.2
langchain-together==0.2.0
langsmith==0.1.147
magic-filter==1.0.12
mando==0.7.1
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==3.0.2
marshmallow==3.23.1
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.4.1
openai==1.55.2
orjson==3.10.12
packaging==24.2
pandas==2.2.3
pbr==6.1.0
pdfkit==1.0.0
pillow==10.4.0
platformdirs==4.3.6
portalocker==2.10.1
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==5.29.0
pyarrow==18.1.0
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydyf==0.11.0
Pygments==2.18.0
pylint==3.3.1
pypdf==5.1.0
PyPDF2==3.0.1
pyphen==0.15.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.12.1
radon==6.0.1
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.2
stevedore==5.4.0
structlog==24.4.0
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
together==1.3.5
tokenizers==0.20.3
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typer==0.13.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
weasyprint==63.0
webencodings==0.5.1
yarl==1.18.0
zopfli==0.2.3.post1
